benchdnn: utils: move memory functionality into a separate file
