+++ tests/benchdnn/binary/binary.cpp
@@ -23,6 +23,7 @@
 #include "oneapi/dnnl/dnnl.h"
 
 #include "utils/fill.hpp"
+#include "utils/memory.hpp"
 #include "utils/parallel.hpp"
 
 #include "dnn_types.hpp"
