+++ tests/benchdnn/common.hpp
@@ -164,10 +164,6 @@ void parse_result(res_t &res, const char *pstr);
 /* misc */
 void init_fp_mode();
 
-void *zmalloc(size_t size, size_t align);
-void zfree(void *ptr);
-void set_zmalloc_max_expected_size(size_t size);
-
 bool str2bool(const char *str);
 const char *bool2str(bool value);
