+++ tests/benchdnn/deconv/deconv.cpp
@@ -20,6 +20,7 @@
 #include <utility>
 
 #include "utils/fill.hpp"
+#include "utils/memory.hpp"
 #include "utils/parallel.hpp"
 
 #include "dnnl_common.hpp"
