+++ tests/benchdnn/reduction/reduction.cpp
@@ -20,6 +20,7 @@
 #include <sstream>
 
 #include "utils/fill.hpp"
+#include "utils/memory.hpp"
 #include "utils/parallel.hpp"
 
 #include "dnnl_common.hpp"
