+++ tests/benchdnn/rnn/rnn.hpp
@@ -28,6 +28,7 @@
 #include "dnn_types.hpp"
 #include "dnnl_common.hpp"
 #include "dnnl_debug.hpp"
+#include "utils/memory.hpp"
 #include "utils/perf_report.hpp"
 #include "utils/settings.hpp"
