+++ tests/benchdnn/shuffle/shuffle.cpp
@@ -20,6 +20,7 @@
 
 #include "oneapi/dnnl/dnnl.h"
 
+#include "utils/memory.hpp"
 #include "utils/parallel.hpp"
 
 #include "dnnl_common.hpp"
